+++ h5netcdf/attrs.py
@@ -26,12 +26,41 @@ class Attributes(MutableMapping):
 
         if key in _HIDDEN_ATTRS:
             raise KeyError(key)
+
+        # get original attribute via h5py low level api
+        # see https://github.com/h5py/h5py/issues/2045
+        attr = self._h5attrs.get_id(key)
+
         # see https://github.com/h5netcdf/h5netcdf/issues/94 for details
         if isinstance(self._h5attrs[key], h5py.Empty):
             string_info = h5py.check_string_dtype(self._h5attrs[key].dtype)
             if string_info and string_info.length == 1:
                 return b""
-        return self._h5attrs[key]
+
+        # transform to 1d array in any case for easy iteration
+        # see https://github.com/h5netcdf/h5netcdf/issues/116
+        output = np.atleast_1d(self._h5attrs[key])
+
+        # string decoding subtleties
+        # vlen strings are already decoded -> only decode fixed length strings
+        # see https://github.com/h5netcdf/h5netcdf/issues/116
+        # netcdf4-python returns string arrays as lists, we do as well
+        string_info = h5py.check_string_dtype(attr.dtype)
+        if string_info is not None:
+            # do not decode "S1"-type char arrays, as they are actually wanted as bytes
+            # see https://github.com/Unidata/netcdf4-python/issues/271
+            if string_info.length is not None and string_info.length > 1:
+                encoding = string_info.encoding
+                output = [b.decode(encoding, "surrogateescape") for b in output.flat]
+            else:
+                output = list(output)
+
+        # return item if single element list/array
+        # see https://github.com/h5netcdf/h5netcdf/issues/116
+        if len(output) == 1:
+            return output[0]
+
+        return output
 
     def __setitem__(self, key, value):
         if key in _HIDDEN_ATTRS: